Validaciones POST y mensajes de backend

+++ database/migrations/3.2022_10_07_121517_create_producto_table.php
@@ -19,8 +19,8 @@ return new class extends Migration
             $table->string('nombre')->unique();;
             $table->binary('imagen')->nullable();
             $table->string('descripcion');
-            $table->integer('precio');
-            $table->integer('stock');
+            $table->unsignedBigInteger('precio');
+            $table->unsignedBigInteger('stock');
             $table->timestamps();
 
             $table->unsignedBigInteger('categoria_id');
